更新場景gizmo

+++ HeroFishing/Assets/Scripts/GameBridge/Pack/Matchgame_ToClient/UPDATESCENE_TOCLIENT.cs
@@ -14,7 +14,7 @@ namespace HeroFishing.Socket.Matchgame {
     }
     public class Spawn {
         public int RouteJsonID { get; private set; }// 路徑JsonID
-        public float SpawnTime { get; private set; }// 在遊戲時間第X秒時被產生的
+        public double SpawnTime { get; private set; }// 在遊戲時間第X秒時被產生的
         public bool IsBoss { get; private set; }// 是否為Boss生怪
 
         public Monster[] Monsters { get; private set; }// 怪物清單
